(chore) frontend code packages re-organization
Given that there is going to be a lot of code functionality to be
shared between the Vue web application and the checkout20 component,
the idea is to have the following structure:

* Backend
  - hub20: an independent python package that can be downloaded from
    PyPI and also a docker image

* Frontend
  - vue-hub20: a javascript package that contains all the library
    logics and vue components that can be shared
  - hub20_web: a javascript package that contain the Vue.js web
    application that provides the functionality for the
    users/operators of the hub (wallet functions and set up of payment
    gateway). This package will depend on vue-hub20
  - checkout20: a javascript package that provides the checkout widget
    to be used by third-party sites that want to integrate with hub20
    stores

Eventually, each of these systems will be in their own git
repositories and have their own independent release-cycle and this
development repository will just use these repositories as git
submodules. They are just being merged here until there is no clean
separation of the packages and a proper CI process for all of them.