simplified event vuex module
Instead of trying to handle all events on that module and then having
components dealing with different combinations before hand, the idea
is to get the event store with a single websocket initialized, and an
action that can define what is the onmessage handler. This way, each
application can define its own event handlers and run the vuex
mutations that they are interested in/responsible for.

This commit introduces this concept by taking the event of new block
creation from the backend and updating the state of the vuex server
module, which is then used by the PaymentRouteBlockchainTimer component.